python/requirements.txt
asgi-correlation-id==4.3.1
fastapi==0.111.1
pydantic==2.8.2
tenacity==8.5.0
uvicorn==0.30.3
requests==2.32.3
pytest-asyncio==0.24.0
httpx==0.28.1
